More instructions and bug fixes

+++ GameVoid/include/Registers.h
@@ -53,8 +53,8 @@ public:
 	void setBC(const WORD &value);
 	void setDE(const WORD &value);
 	void setHL(const WORD &value);
-	void addPC(const BYTE &value);
-	void addSP(const BYTE &value);
+	void addPC(const int &value);
+	void addSP(const int &value);
 	void setSP(const WORD &value);
 	void setPC(const WORD &value);
